Fix counter method name conflict with ActiveRecord
Both Redis::Objects and ActiveRecord::Associations define increment_counter and decrement_counter available to the base
class.  In practice this leads to ActiveRecord counter caches being unusable if Redis::Objects are included on the
class.  As much as counter caches suck, there may be cases where it's necessary to use both.